Extract some utility functions from search
